chore: check for circular references on build (#963)
