+++ package.json
@@ -28,6 +28,7 @@
     "test": "npm run test:unit",
     "test:unit": "vitest --environment jsdom",
     "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
+    "circular-check": "tsc --showConfig -p ./tsconfig.app.json > temp.json && madge --ts-config ./temp.json --warning --circular --extensions ts,vue,js ./src",
     "theme:convert": "cd tools; node convertTheme.js"
   },
   "main": "index.js",
@@ -88,6 +89,7 @@
     "eslint-plugin-vue": "^9.7.0",
     "husky": "^7.0.4",
     "jsdom": "^20.0.2",
+    "madge": "^5.0.1",
     "mockdate": "^3.0.5",
     "monaco-vscode-textmate-theme-converter": "^0.1.7",
     "sass": "~1.32",
