+++ src/store/config/state.ts
@@ -11,24 +11,6 @@ export const defaultState = (): ConfigState => {
       endpoints: [],
       blacklist: [],
       hosted: false,
-      locales: [
-        { name: 'English', code: 'en' },
-        { name: 'Čeština', code: 'cs' },
-        { name: '简体中文', code: 'zh-CN' },
-        { name: '繁體中文', code: 'zh-HK' },
-        { name: 'Deutsch', code: 'de' },
-        { name: 'Español', code: 'es' },
-        { name: 'Français', code: 'fr' },
-        { name: 'Italiano', code: 'it' },
-        { name: 'Magyar', code: 'hu' },
-        { name: 'Nederlands', code: 'nl' },
-        { name: 'Português', code: 'pt' },
-        { name: 'Русский', code: 'ru' },
-        { name: 'Українська', code: 'uk' },
-        { name: '한국어', code: 'ko' },
-        { name: 'العربية', code: 'ar' },
-        { name: '日本語', code: 'ja' }
-      ],
       themePresets: []
     },
     instances: [],
