+++ src/store/files/types.ts
@@ -1,4 +1,7 @@
+import { KlipperFileMeta, Thumbnail } from './types.metadata'
 import { HistoryItem } from '@/store/history/types'
+
+export type { KlipperFileMeta, Thumbnail }
 
 export interface FilesState {
   uploads: FilesUpload[];
@@ -48,27 +51,6 @@ export interface KlipperFile {
   job_id?: string | null;
 }
 
-export interface KlipperFileMeta {
-  chamber_temp?: number;
-  estimated_time?: number;
-  filament_name?: string;
-  filament_total?: number;
-  filament_type?: string;
-  filament_weight_total?: number;
-  first_layer_bed_temp?: number;
-  first_layer_extr_temp?: number;
-  first_layer_height?: number;
-  gcode_end_byte?: number;
-  gcode_start_byte?: number;
-  layer_count?: number;
-  layer_height?: number;
-  nozzle_diameter?: number;
-  object_height?: number;
-  slicer?: string;
-  slicer_version?: string;
-  thumbnails?: Thumbnail[];
-}
-
 export interface AppFileWithMeta extends AppFile, KlipperFileMeta {
   history: HistoryItem;
 }
@@ -79,15 +61,6 @@ export interface AppDirectory {
   name?: string;
   dirname: string;
   modified: number | null;
-  size: number;
-}
-
-export interface Thumbnail {
-  data: string;
-  relative_path: string;
-  absolute_path?: string;
-  height: number;
-  width: number;
   size: number;
 }
